backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
Werkzeug==3.0.1
bcrypt==4.1.2
google-generativeai==0.3.2
PyPDF2==3.0.1
python-multipart==0.0.6
Pillow==10.1.0
requests==2.31.0
gunicorn==21.2.0